src/test/java/ru/job4j/collection/PhoneDictionaryTest.java
package ru.job4j.collection;

import org.junit.Test;

import java.util.ArrayList;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.*;

public class PhoneDictionaryTest {

    @Test
    public void whenFindBySurName() {
        var phone = new PhoneDictionary();
        phone.add(new Person("Danila", "Ryzhov", "93612495", " Kuzneck"));
        ArrayList<Person> persons = phone.find("zhov");
        assertThat(persons.get(0).getSurname(), is("Ryzhov"));
    }

    @Test
    public void whenFindByName() {
        var phones = new PhoneDictionary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        ArrayList<Person> persons = phones.find("Petr");
        assertThat(persons.get(0).getSurname(), is("Arsentev"));
    }

    @Test
    public void whenFindByNamePredicate() {
        var phones = new PhoneDictionary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        ArrayList<Person> persons = phones.find("Petr");
        assertThat(persons.get(0).getSurname(), is("Arsentev"));
    }
}